[BERDA] Restructured the files
