+++ examscheduler.client/src/App.jsx
@@ -1,8 +1,8 @@
 ﻿import React from 'react';
 import { BrowserRouter as Router, Route, Routes } from 'react-router-dom';
-import RegisterPage from './RegisterPage';
-import LoginPage from './LoginPage';
-import DashboardPage from './DashboardPage';
+import RegisterPage from './Pages/RegisterPage';
+import LoginPage from './Pages/LoginPage';
+import DashboardPage from './Pages/DashboardPage';
 
 const App = () => {
     return (
